requirements.txt
numpy~=1.17.4
pandas~=0.25.3
opencv-python-headless
torch~=1.3.1+cpu
torchvision~=0.4.2+cpu
scipy~=1.3.3
tqdm~=4.40.2
Pillow~=6.2.1
opencv-python~=4.1.2.30
matplotlib~=3.1.2
scikit-image~=0.16.2
DPOD~=0.1
setuptools~=40.8.0